refactor code for html generation
