+++ dotnet-interactive.sln
@@ -35,6 +35,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DotNet.Interactiv
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "dotnet-interactive.Tests", "dotnet-interactive.Tests\dotnet-interactive.Tests.csproj", "{5AA74BE6-5D52-43E5-8654-5F5640F65B8F}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "XPlot.DotNet.Interactive.KernelExtensions", "XPlot.DotNet.Interactive.KernelExtensions\XPlot.DotNet.Interactive.KernelExtensions.csproj", "{F3479E07-528D-42FD-829B-5A801A1C8562}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -213,6 +215,18 @@ Global
 		{5AA74BE6-5D52-43E5-8654-5F5640F65B8F}.Release|x64.Build.0 = Release|Any CPU
 		{5AA74BE6-5D52-43E5-8654-5F5640F65B8F}.Release|x86.ActiveCfg = Release|Any CPU
 		{5AA74BE6-5D52-43E5-8654-5F5640F65B8F}.Release|x86.Build.0 = Release|Any CPU
+		{F3479E07-528D-42FD-829B-5A801A1C8562}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F3479E07-528D-42FD-829B-5A801A1C8562}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F3479E07-528D-42FD-829B-5A801A1C8562}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{F3479E07-528D-42FD-829B-5A801A1C8562}.Debug|x64.Build.0 = Debug|Any CPU
+		{F3479E07-528D-42FD-829B-5A801A1C8562}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{F3479E07-528D-42FD-829B-5A801A1C8562}.Debug|x86.Build.0 = Debug|Any CPU
+		{F3479E07-528D-42FD-829B-5A801A1C8562}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F3479E07-528D-42FD-829B-5A801A1C8562}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F3479E07-528D-42FD-829B-5A801A1C8562}.Release|x64.ActiveCfg = Release|Any CPU
+		{F3479E07-528D-42FD-829B-5A801A1C8562}.Release|x64.Build.0 = Release|Any CPU
+		{F3479E07-528D-42FD-829B-5A801A1C8562}.Release|x86.ActiveCfg = Release|Any CPU
+		{F3479E07-528D-42FD-829B-5A801A1C8562}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -232,6 +246,7 @@ Global
 		{8BF966AA-96BE-4E55-AD0F-8AEC90B0254F} = {11BA3480-4584-435C-BA9A-8C554DB60E9F}
 		{73136CC6-D968-4D2C-AD93-B23A78B2C7DF} = {11BA3480-4584-435C-BA9A-8C554DB60E9F}
 		{5AA74BE6-5D52-43E5-8654-5F5640F65B8F} = {11BA3480-4584-435C-BA9A-8C554DB60E9F}
+		{F3479E07-528D-42FD-829B-5A801A1C8562} = {B95A8485-8C53-4F56-B0CE-19C0726B5805}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6D05A9AF-CFFB-4187-8599-574387B76727}
